- Contrib plugins: Use %zu when formatting size_t. Remove "static_cast<unsigned long>()" calls.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@13103 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/ThreadSearch/ThreadSearchLoggerList.cpp
@@ -460,7 +460,7 @@ void ThreadSearchLoggerList::OnSearchBegin(const ThreadSearchFindData& findData)
 
 void ThreadSearchLoggerList::OnSearchEnd()
 {
-    wxString message = wxString::Format(_("%lu matches found."), static_cast<unsigned long>(m_TotalLinesFound));
+    wxString message = wxString::Format(_("%zu matches found."), m_TotalLinesFound);
     long index = m_pListLog->GetItemCount();
     m_pListLog->InsertItem(index, _("=> Search complete. "));
     m_pListLog->SetItem(index, ListColumns::Text, message);
